Add french dictionary and update test
